+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/viewmodels/BaseSheetViewModel.kt
@@ -12,11 +12,9 @@ import com.stripe.android.link.ui.inline.InlineSignupViewState
 import com.stripe.android.link.ui.inline.LinkSignupMode
 import com.stripe.android.link.ui.inline.UserInput
 import com.stripe.android.lpmfoundations.luxe.SupportedPaymentMethod
-import com.stripe.android.lpmfoundations.luxe.isSaveForFutureUseValueChangeable
 import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadata
 import com.stripe.android.lpmfoundations.paymentmethod.UiDefinitionFactory
 import com.stripe.android.model.CardBrand
-import com.stripe.android.model.PaymentIntent
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodCode
 import com.stripe.android.model.PaymentMethodCreateParams
@@ -41,7 +39,6 @@ import com.stripe.android.paymentsheet.navigation.PaymentSheetScreen
 import com.stripe.android.paymentsheet.navigation.PaymentSheetScreen.AddAnotherPaymentMethod
 import com.stripe.android.paymentsheet.navigation.PaymentSheetScreen.AddFirstPaymentMethod
 import com.stripe.android.paymentsheet.paymentdatacollection.FormArguments
-import com.stripe.android.paymentsheet.paymentdatacollection.ach.USBankAccountFormArguments
 import com.stripe.android.paymentsheet.repositories.CustomerRepository
 import com.stripe.android.paymentsheet.state.CustomerState
 import com.stripe.android.paymentsheet.state.GooglePayState
@@ -808,40 +805,6 @@ internal abstract class BaseSheetViewModel(
 
     fun reportCardNumberCompleted() {
         eventReporter.onCardNumberCompleted()
-    }
-
-    fun usBankAccountFormArguments(selectedPaymentMethodCode: String): USBankAccountFormArguments {
-        val isSaveForFutureUseValueChangeable = paymentMethodMetadata.value?.let {
-            isSaveForFutureUseValueChangeable(
-                code = selectedPaymentMethodCode,
-                metadata = it,
-            )
-        } ?: false
-        val instantDebits = selectedPaymentMethodCode == PaymentMethod.Type.Link.code
-        val initializationMode = (this as? PaymentSheetViewModel)
-            ?.args
-            ?.initializationMode
-        val onBehalfOf = (initializationMode as? PaymentSheet.InitializationMode.DeferredIntent)
-            ?.intentConfiguration
-            ?.onBehalfOf
-        val stripeIntent = paymentMethodMetadata.value?.stripeIntent
-        return USBankAccountFormArguments(
-            showCheckbox = isSaveForFutureUseValueChangeable,
-            instantDebits = instantDebits,
-            onBehalfOf = onBehalfOf,
-            isCompleteFlow = this is PaymentSheetViewModel,
-            isPaymentFlow = stripeIntent is PaymentIntent,
-            stripeIntentId = stripeIntent?.id,
-            clientSecret = stripeIntent?.clientSecret,
-            shippingDetails = config.shippingDetails,
-            draftPaymentSelection = newPaymentSelection?.paymentSelection,
-            onMandateTextChanged = ::updateMandateText,
-            onConfirmUSBankAccount = ::handleConfirmUSBankAccount,
-            onCollectBankAccountResult = null,
-            onUpdatePrimaryButtonUIState = ::updateCustomPrimaryButtonUiState,
-            onUpdatePrimaryButtonState = ::updatePrimaryButtonState,
-            onError = ::onError
-        )
     }
 
     fun onFormFieldValuesChanged(formValues: FormFieldValues?, selectedPaymentMethodCode: String) {
